Вынесла функцию лайков в модуль
